# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Credit Risk Workflow API
  version: 1.0.0
  description: |
    REST API for Credit Risk Workflow system, including credit requests, workflow transitions, documents, notifications, and PDF download.
servers:
  - url: http://localhost:8001/api/
paths:
...
  /documents/{id}/new_version/:
    post:
      summary: Create new version of a document
      tags: [Document]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                description:
                  type: string
      responses:
        '200':
          description: New version created
        '400':
          description: Permission denied
  /documents/{id}/finalize/:
    post:
      summary: Finalize (publish) a document
      tags: [Document]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Finalized
        '400':
          description: Permission denied
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []